update to handle null schemaFiles
git-svn-id: http://svn.codehaus.org/geoserver/trunk@3018 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 909a5ed5996829022a6ea7ac9d41fa6ac32c5754 [formerly 856faa9cdc68f83b073d716b5cff3c1093c42d03]
Former-commit-id: 5a44ca81d13e9f9f77d354726463ff6bc98119a5

+++ geoserver/src/org/vfny/geoserver/responses/wfs/DescribeResponse.java
@@ -327,7 +327,13 @@ public class DescribeResponse implements Response {
 
                 File schemaFile = meta.getSchemaFile();
                    try {
-                       if (schemaFile.exists() && schemaFile.canRead()) {
+                       //Hack here, schemaFile should not be null, but it is
+                       //when a fType is first created, since we only add the 
+                       //schemaFile param to dto on a load.  This should be
+                       //fixed, maybe even have the schema file persist, or at
+                       //the very least be present right after creation.
+                       if (schemaFile != null &&
+                            schemaFile.exists() && schemaFile.canRead()) {
                            generatedType = writeFile(schemaFile);
                        } else {
                            FeatureType ft2 = meta.getFeatureType();
